Behat green for document-management feature against API Context

+++ app/Http/Controllers/Api/DocumentController.php
@@ -97,13 +97,28 @@ class DocumentController extends Controller {
 	/**
 	 * Remove the specified Document.
 	 *
-	 * @param  string $documentName
+	 * @param string $documentName
+     * @param string $path
 	 * @return Response
 	 */
-	public function destroy($documentName)
+	public function destroy($documentName, $path=null)
 	{
-        $this->documentRepository->delete($documentName);
-        return new JsonResponse(array("message" => "Destroyed : {$documentName}"));
+        if ($path)
+        {
+            $document = $this->documentRepository->fetch($documentName);
+            if (!$document instanceof DataDocument)
+            {
+                return new JsonResponse(array("message" => "Document '{$documentName}' not found"), 404);
+            }
+            $document->unsetPath($path);
+            $this->documentRepository->store($document);
+            return new JsonResponse(array("message" => "Deleted : {$documentName} : {$path}"));
+        }
+        else
+        {
+            $this->documentRepository->delete($documentName);
+            return new JsonResponse(array("message" => "Deleted : {$documentName}"));
+        }
 	}
 
 }
